Movie의 할인 금액 계산을 상속에서 합성으로 변경
