+++ src/jooom/chapter05/DiscountPolicy.java
@@ -1,0 +1,5 @@
+package jooom.chapter05;
+
+public abstract class DiscountPolicy {
+    abstract protected Money calculateDiscountAmount();
+}
