--- src/jooom/chapter05/NoneDiscountMovie.java
@@ -1,15 +1,0 @@
-package jooom.chapter05;
-
-import java.time.Duration;
-import java.util.List;
-
-public class NoneDiscountMovie extends Movie{
-    public NoneDiscountMovie(String title, Duration runningTime, Money fee, List<DiscountCondition> discountConditions) {
-        super(title, runningTime, fee, discountConditions);
-    }
-
-    @Override
-    protected Money calculateDiscountAmount() {
-        return Money.ZERO;
-    }
-}
