+++ src/jooom/chapter05/discount/AmountDiscountPolicy.java
@@ -1,0 +1,17 @@
+package jooom.chapter05.discount;
+
+import jooom.chapter05.DiscountPolicy;
+import jooom.chapter05.Money;
+
+public class AmountDiscountPolicy extends DiscountPolicy {
+    private Money discountAmount;
+
+    public AmountDiscountPolicy(Money discountAmount) {
+        this.discountAmount = discountAmount;
+    }
+
+    @Override
+    protected Money calculateDiscountAmount() {
+        return discountAmount;
+    }
+}
