--- src/jooom/chapter05/discount/policy/AmountDiscountPolicy.java
@@ -1,19 +1,0 @@
-package jooom.chapter05.discount.policy;
-
-import jooom.chapter05.Money;
-import jooom.chapter05.Screening;
-import jooom.chapter05.discount.condition.DiscountCondition;
-
-public class AmountDiscountPolicy extends DiscountPolicy {
-    private Money discountAmount;
-
-    public AmountDiscountPolicy(Money discountAmount, DiscountCondition... conditions) {
-        super(conditions);
-        this.discountAmount = discountAmount;
-    }
-
-    @Override
-    protected Money getDiscountAmount(Screening screening) {
-        return discountAmount;
-    }
-}
