--- src/jooom/chapter05/discount/policy/DiscountPolicy.java
@@ -1,28 +1,0 @@
-package jooom.chapter05.discount.policy;
-
-import jooom.chapter05.DiscountCondition;
-import jooom.chapter05.Screening;
-import jooom.chapter05.Money;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-public abstract class DiscountPolicy {
-    private List<DiscountCondition> conditions = new ArrayList<>();
-
-    public DiscountPolicy(DiscountCondition... conditions){
-        this.conditions = Arrays.asList(conditions);
-    }
-
-    public Money calculateDiscountAmount(Screening screening){
-        for (DiscountCondition condition : conditions){
-            if (condition.isSatisfiedBy(screening)){
-                return getDiscountAmount(screening);
-            }
-        }
-        return Money.ZERO;
-    }
-
-    abstract protected Money getDiscountAmount(Screening screening);
-}
